--- backend/infrahub/message_bus/messages/proposed_change/request_proposedchange_runtests.py
@@ -1,5 +1,0 @@
-from .base_with_diff import BaseProposedChangeWithDiffMessage
-
-
-class RequestProposedChangeRunTests(BaseProposedChangeWithDiffMessage):
-    """Sent trigger to run tests (smoke, units, integrations) for a proposed change."""
