+++ backend/infrahub/message_bus/operations/__init__.py
@@ -42,7 +42,6 @@ COMMAND_MAP = {
     "request.artifact_definition.check": requests.artifact_definition.check,
     "request.proposed_change.pipeline": requests.proposed_change.pipeline,
     "request.proposed_change.refresh_artifacts": requests.proposed_change.refresh_artifacts,
-    "request.proposed_change.run_tests": requests.proposed_change.run_tests,
     "request.repository.checks": requests.repository.checks,
     "request.repository.user_checks": requests.repository.user_checks,
     "send.echo.request": send.echo.request,
