+++ backend/tests/integration/message_bus/operations/request/test_proposed_change.py
@@ -153,13 +153,11 @@ async def test_run_pipeline_validate_requested_jobs(
 
         assert sorted(bus_pre_data_changes.seen_routing_keys) == [
             "request.proposed_change.refresh_artifacts",
-            "request.proposed_change.run_tests",
             "request.repository.user_checks",
         ]
 
         assert sorted(bus_post_data_changes.seen_routing_keys) == [
             "request.proposed_change.refresh_artifacts",
-            "request.proposed_change.run_tests",
             "request.repository.user_checks",
             "schema.validator.path",
         ]
