+++ docs/docs/reference/message-bus-events.mdx
@@ -554,24 +554,6 @@ For more detailed explanations on how to use these events within Infrahub, see t
 | **destination_branch** | The destination branch of the proposed change | string | None |
 | **branch_diff** | The calculated diff between the two branches | N/A | None |
 <!-- vale on -->
-<!-- vale off -->
-#### Event request.proposed_change.run_tests
-<!-- vale on -->
-
-**Description**: Sent trigger to run tests (smoke, units, integrations) for a proposed change.
-
-**Priority**: 3
-
-<!-- vale off -->
-| Key | Description | Type | Default Value |
-|-----|-------------|------|---------------|
-| **meta** | Meta properties for the message | N/A | None |
-| **proposed_change** | The unique ID of the Proposed Change | string | None |
-| **source_branch** | The source branch of the proposed change | string | None |
-| **source_branch_sync_with_git** | Indicates if the source branch should sync with git | boolean | None |
-| **destination_branch** | The destination branch of the proposed change | string | None |
-| **branch_diff** | The calculated diff between the two branches | N/A | None |
-<!-- vale on -->
 
 <!-- vale off -->
 ### Request Repository
@@ -1211,25 +1193,6 @@ For more detailed explanations on how to use these events within Infrahub, see t
 <!-- vale on -->
 
 **Description**: Sent trigger the refresh of artifacts that are impacted by the proposed change.
-
-**Priority**: 3
-
-
-<!-- vale off -->
-| Key | Description | Type | Default Value |
-|-----|-------------|------|---------------|
-| **meta** | Meta properties for the message | N/A | None |
-| **proposed_change** | The unique ID of the Proposed Change | string | None |
-| **source_branch** | The source branch of the proposed change | string | None |
-| **source_branch_sync_with_git** | Indicates if the source branch should sync with git | boolean | None |
-| **destination_branch** | The destination branch of the proposed change | string | None |
-| **branch_diff** | The calculated diff between the two branches | N/A | None |
-<!-- vale on -->
-<!-- vale off -->
-#### Event request.proposed_change.run_tests
-<!-- vale on -->
-
-**Description**: Sent trigger to run tests (smoke, units, integrations) for a proposed change.
 
 **Priority**: 3
